// app/Models/Admission/admission.ts
export class Admission {
  id: number;
  admId:number;
  firstName: string;
  idNum:string;
  parentId: number;
  fatherFullName: string;
  fatherFirstName: string;
  fatherSecondName: string;
  fatherFamilyName: string;
  studNo: number;
  schoolId: number;
  sectionId: number;
  entryDate: Date;
  nationalityId: number;
  religionId: number;
  birthDate: Date;
  genderId: number;
  genderName: string;
  yearId: number;
  classId: number;
  classPrice: number;
  className: string;
  classSeqId: number;
  classSeqName: string;
  descountType: number;
  descountValue: number;
  tourId: number;
  tourTypeId: number;
  tourName: string;
  tourTypeName: string;
  tourPrice: number;
  busId: number;
  approvedId: number;
  approvedDate: Date;
  studentBrotherSeq: number;
  brotherDescountType: number;
  busNote: string;
  note: string;
  totalPrice: number;
  sectionName:string;
  status:string;
  SchoolId:number;
  studFullName:string;
  parentName:string;
  studId:string;
  fatherMobail:number;
  motherMobail:number;
  studName:string;
  classSeq:number;
  studentId:number;
  result:any=[];
}